+++ mssql-managed-instance/test-case/locals.tf
@@ -10,17 +10,38 @@ locals {
     {
       name                         = local.naming.mi_1
       location                     = local.location
-      resource_group_name          = local.naming.rg
+      resource_group_name          = azurerm_resource_group.rg.name
       administrator_login          = "useradmin"
       administrator_login_password = "Password1234"
       license_type                 = "LicenseIncluded"
       sku_name                     = "GP_Gen5"
       storage_size_in_gb           = 32
       vcores                       = 4
-      subnet_id                    = data.azurerm_subnet.example.id
+      subnet_id                    = azurerm_subnet.example.id
       identity = {
         type = "SystemAssigned"
       }
+
+      private_endpoint = [
+        {
+          name                          = "${local.naming.mi_1}-PE01"
+          subnet_id                     = azurerm_subnet.subnet.id
+          custom_network_interface_name = "${local.naming.mi_1}-PE01.nic"
+          private_service_connection = [
+            {
+              name                 = "${local.naming.mi_1}-PE01-connection"
+              is_manual_connection = false
+              subresource_names    = ["managedInstance"]
+            }
+          ]
+          private_dns_zone_group = {
+            name = azurerm_private_dns_zone.dns.name
+            private_dns_zone_ids = [
+              azurerm_private_dns_zone.dns.id
+            ]
+          }
+        }
+      ]
 
       monitoring = [
         {