+++ mssql-managed-instance/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Release v1.4.0
+
+## Provider & Terraform Upgrade
+- Azurerm provider: 3.96.0 (#15)
+- Terraform: 1.7.5 (#15)
+
+## Enhancements
+- Add private endpoint (#15)
+   
 ## Release v1.3.1
 
 ## Enhancements